Update the Delete Page https://docs.microsoft.com/en-us/aspnet/core/data/ef-mvc/crud

+++ ContosoUniversity/Controllers/StudentsController.cs
@@ -105,20 +105,15 @@ namespace ContosoUniversity.Controllers
 		}
 
 		// GET: Students/Delete/5
-		public async Task<IActionResult> Delete(int? id)
+		public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
 		{
-			if (id == null)
+			if (id == null) return NotFound();
+			var student = await _context.Students.AsNoTracking().SingleOrDefaultAsync(m => m.ID == id);
+			if (student == null) return NotFound();
+			if (saveChangesError.GetValueOrDefault())
 			{
-				return NotFound();
+				ViewData["ErrorMessage"] = "Delete failed. Try again, and if the problem persists see your system administrator.";
 			}
-
-			var student = await _context.Students
-					.SingleOrDefaultAsync(m => m.ID == id);
-			if (student == null)
-			{
-				return NotFound();
-			}
-
 			return View(student);
 		}
 
@@ -127,10 +122,19 @@ namespace ContosoUniversity.Controllers
 		[ValidateAntiForgeryToken]
 		public async Task<IActionResult> DeleteConfirmed(int id)
 		{
-			var student = await _context.Students.SingleOrDefaultAsync(m => m.ID == id);
-			_context.Students.Remove(student);
-			await _context.SaveChangesAsync();
-			return RedirectToAction("Index");
+			var student = await _context.Students.AsNoTracking().SingleOrDefaultAsync(m => m.ID == id);
+			if (student == null) return RedirectToAction("Index");
+			try
+			{
+				_context.Students.Remove(student);
+				await _context.SaveChangesAsync();
+				return RedirectToAction("Index");
+			}
+			catch (DbUpdateException)
+			{
+				return RedirectToAction("Delete", new { id = id, saveChangesError = true });
+			}
+
 		}
 
 		private bool StudentExists(int id)
